Split the citation extractor and its tests under the transform structure caps
